+++ docs/Whats-New.md
@@ -405,6 +405,7 @@ HideShakeEffects=false           ; boolean
 :open:
 
 #### New:
+- [Jumpjet vehicle carryall](New-or-Enhanced-Logics.md#jumpjet-vehicle-carryall) (by BlackgamerzVN)
 - [Allow using waypoints, area guard and attack move with aircraft](Fixed-or-Improved-Logics.md#extended-aircraft-missions) (by CrimRecya)
 - [Enhanced Straight trajectory](New-or-Enhanced-Logics.md#straight-trajectory) (by CrimRecya)
 - [Enable building production queue](User-Interface.md#building-production-queue) (by CrimRecya)